requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
typing
python-dotenv
psycopg2